back-end/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
Werkzeug==2.0.1
python-dotenv==0.19.0
Flask-Testing==0.8.1  # Added for testing
python-dotenv==0.19.0